Lazy plugs (#596)
* Manifest caching and lazy loading of plug workers
* Fixes #546 Plug unloading after time out

+++ plugos/sandbox.ts
@@ -9,26 +9,38 @@ export type SandboxFactory<HookT> = (plug: Plug<HookT>) => Sandbox<HookT>;
  * Effectively this wraps a web worker, the reason to have this split from Plugs is to allow plugs to manage multiple sandboxes, e.g. for performance in the future
  */
 export class Sandbox<HookT> {
-  private worker: Worker;
+  private worker?: Worker;
   private reqId = 0;
   private outstandingInvocations = new Map<
     number,
     { resolve: (result: any) => void; reject: (e: any) => void }
   >();
 
-  public ready: Promise<void>;
+  // public ready: Promise<void>;
   public manifest?: Manifest<HookT>;
 
   constructor(
     readonly plug: Plug<HookT>,
-    workerOptions = {},
+    private workerOptions = {},
   ) {
-    this.worker = new Worker(plug.workerUrl, {
-      ...workerOptions,
+  }
+
+  /**
+   * Should only invoked lazily (either by invoke, or by a ManifestCache to load the manifest)
+   */
+  init(): Promise<void> {
+    console.log("Booting up worker for", this.plug.name);
+    if (this.worker) {
+      // Should not happen
+      console.warn("Double init of sandbox");
+    }
+    this.worker = new Worker(this.plug.workerUrl, {
+      ...this.workerOptions,
       type: "module",
     });
-    this.ready = new Promise((resolve) => {
-      this.worker.onmessage = (ev) => {
+
+    return new Promise((resolve) => {
+      this.worker!.onmessage = (ev) => {
         if (ev.data.type === "manifest") {
           this.manifest = ev.data.manifest;
           resolve();
@@ -46,14 +58,14 @@ export class Sandbox<HookT> {
         try {
           const result = await this.plug.syscall(data.name!, data.args!);
 
-          this.worker.postMessage({
+          this.worker!.postMessage({
             type: "sysr",
             id: data.id,
             result: result,
           } as WorkerMessage);
         } catch (e: any) {
           // console.error("Syscall fail", e);
-          this.worker.postMessage({
+          this.worker!.postMessage({
             type: "sysr",
             id: data.id,
             error: e.message,
@@ -76,9 +88,13 @@ export class Sandbox<HookT> {
     }
   }
 
-  invoke(name: string, args: any[]): Promise<any> {
+  async invoke(name: string, args: any[]): Promise<any> {
+    if (!this.worker) {
+      // Lazy initialization
+      await this.init();
+    }
     this.reqId++;
-    this.worker.postMessage({
+    this.worker!.postMessage({
       type: "inv",
       id: this.reqId,
       name,
@@ -90,6 +106,9 @@ export class Sandbox<HookT> {
   }
 
   stop() {
-    this.worker.terminate();
+    if (this.worker) {
+      this.worker.terminate();
+      this.worker = undefined;
+    }
   }
 }
